requirements.txt
##  Using pip freeze ## pip install -r requirements_linux.txt ## or rename this file to requirements.txt and run same command without _linux
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
asttokens==2.4.1
attrs==23.2.0
az==0.1.0.dev1
azure-ai-documentintelligence==1.0.0b1
azure-ai-formrecognizer==3.3.2
azure-ai-ml==1.13.0
azure-cognitiveservices-speech==1.34.0
azure-common==1.1.28
azure-core==1.30.0
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
decorator==5.1.1
deprecation==2.1.0
elevenlabs==0.2.26
executing==2.0.1
fastjsonschema==2.19.1
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.41
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-shell==1.8.0
google-cloud-speech==2.23.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.15.0
google-cloud-vision==3.5.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
google-speech==1.2.0
googleapis-common-protos==1.62.0
googletrans==4.0.0rc1
grpcio==1.60.0
grpcio-status==1.60.0
gTTS==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
img2pdf==0.5.0
importlib-metadata==7.0.1
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mdurl==0.1.2
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
multidict==6.0.5
nbformat==5.9.2
numpy==1.26.4
oauthlib==3.2.2
ocrmypdf==15.4.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdfminer.six==20221105
pigar==2.1.3
pikepdf==8.6.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.3.0
portalocker==2.8.2
prompt-toolkit==3.0.41
proto-plus==1.23.0
protobuf==4.25.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydash==7.0.5
pydeck==0.8.1b0
pyee==11.1.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-ffmpeg==2.0.10
pytz==2024.1
## pywin32==306
PyYAML==6.0.1
referencing==0.33.0
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.18.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
streamlit==1.31.1
strictyaml==1.7.3
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.12
web-cache==1.1.0
websockets==12.0
yarl==1.9.4
zipp==3.17.0
